Fix UI update
